Resolved PR reviews
