+++ semantic-convention-utils/src/main/java/org/hypertrace/semantic/convention/utils/span/SpanSemanticConventionUtils.java
@@ -2,8 +2,10 @@ package org.hypertrace.semantic.convention.utils.span;
 
 import java.util.Map;
 import java.util.Optional;
+import org.apache.avro.reflect.Nullable;
 import org.apache.commons.lang3.StringUtils;
 import org.hypertrace.core.datamodel.AttributeValue;
+import org.hypertrace.core.datamodel.Attributes;
 import org.hypertrace.core.datamodel.Event;
 import org.hypertrace.core.datamodel.shared.SpanAttributeUtils;
 import org.hypertrace.core.semantic.convention.constants.span.OTelSpanSemanticConventions;
@@ -11,6 +13,7 @@ import org.hypertrace.core.semantic.convention.constants.span.OpenTracingSpanSem
 import org.hypertrace.core.span.constants.RawSpanConstants;
 import org.hypertrace.core.span.constants.v1.OCAttribute;
 import org.hypertrace.core.span.constants.v1.OCSpanKind;
+import org.hypertrace.semantic.convention.utils.http.HttpSemanticConventionUtils;
 
 /** Utility to read span attributes */
 public class SpanSemanticConventionUtils {
@@ -143,5 +146,24 @@ public class SpanSemanticConventionUtils {
       return SpanAttributeUtils.getStringAttribute(event, OT_PEER_NAME);
     }
     return null;
+  }
+
+  public static boolean isEmptyAttributesMap(Event event) {
+    return Optional.ofNullable(event.getAttributes()).map(Attributes::getAttributeMap).isEmpty();
+  }
+
+  public static boolean isValueNotNull(AttributeValue attributeValue) {
+    return Optional.ofNullable(attributeValue).map(AttributeValue::getValue).isPresent();
+  }
+
+  @Nullable
+  public static String getStringAttribute(Event event, String attributeKey) {
+    AttributeValue value = SpanAttributeUtils.getAttributeValue(event, attributeKey);
+    return value == null ? null : value.getValue();
+  }
+
+  public static boolean isHttpResponseCookie(String requestHeaderAttributeKey) {
+    return requestHeaderAttributeKey.startsWith(
+        HttpSemanticConventionUtils.RESPONSE_COOKIE_HEADER_PREFIX);
   }
 }
